+++ test/conftest.py
@@ -102,6 +102,14 @@ class Helper:
 
         return result
 
+    def show_alert(self, database, schema, name):
+        cur = self.execute("SHOW ALERTS LIKE {alert_name:lf} IN SCHEMA {schema_name:i}", {
+            "schema_name": SchemaIdent(self.env_prefix, database, schema),
+            "alert_name": Ident(name),
+        })
+
+        return cur.fetchone()
+
     def show_sequence(self, database, schema, name):
         cur = self.execute("SHOW SEQUENCES LIKE {sequence_name:lf} IN SCHEMA {schema_name:i}", {
             "schema_name": SchemaIdent(self.env_prefix, database, schema),